Added Image Processing Code
Added my image processing code that includes Brien's code as well. removeLabel.py current just requires the user to pass in an image to the removeLabel() function and it will return an image with the labels removed. To get it to work on your machine you will just have to change the reading in code on line 171 and the saving code on line 188 to point to the correct location on your computer and also have opencv installed I believe. 

Currently, the function is just called in the main at the bottom but can be changed imported by other python files and called that way as well.